+++ Sources/SwagGenKit/SwaggerExtensions.swift
@@ -154,12 +154,7 @@ extension Schema {
                 break
             }
         case let .group(group):
-            switch group.type {
-            case .any, .one:
-                if group.discriminator != nil {
-                    return self
-                }
-            case .all:
+            if group.discriminator != nil {
                 return self
             }
         default: break