+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixed
 - Fixed inline allOf group generation
+- Fixed property generation when there is only one group schema;  the first group schema type will be used as the type #217
 
 ## 4.3.1